fix: refactor api_clients
